rest/requirements.txt
flask-restplus
pyaml
grpcio
uwsgi
protobuf